requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# case-law-explorer/data_extraction/caselaw/cellar/cellar_extraction.py: 13
SPARQLWrapper == 2.0.0

# case-law-explorer/data_loading/clients/dynamodb.py: 1
# case-law-explorer/data_loading/clients/opensearch.py: 1
# case-law-explorer/definitions/storage_handler.py: 4
boto3 == 1.18.56

# case-law-explorer/data_loading/clients/opensearch.py: 2
# case-law-explorer/definitions/storage_handler.py: 5
botocore == 1.21.56

# case-law-explorer/data_extraction/caselaw/rechtspraak/rechtspraak_metadata_dump_parser.py: 4,5
# case-law-explorer/data_extraction/citations/rechtspraak_citations_extractor_incremental.py: 2
# case-law-explorer/data_transformation/utils.py: 6
lxml == 4.6.3

# case-law-explorer/data_transformation/utils.py: 2
numpy == 1.19.2

# case-law-explorer/data_loading/clients/opensearch.py: 4
opensearch_py == 1.0.0

# case-law-explorer/data_extraction/caselaw/echr/ECHR_metadata_harvester.py: 6
# case-law-explorer/data_extraction/caselaw/legal_intelligence/legal_intelligence_extractor.py: 6
# case-law-explorer/data_extraction/caselaw/rechtspraak/rechtspraak_dump_unzipper.py: 11
# case-law-explorer/data_extraction/citations/rechtspraak_citations_extractor_incremental.py: 9
# case-law-explorer/data_transformation/utils.py: 1
# case-law-explorer/definitions/storage_handler.py: 242
# case-law-explorer/notebooks/ECHR_metadata_harvester.ipynb: 2
# case-law-explorer/notebooks/analysis/data_consistency_checks.ipynb: 1
# case-law-explorer/notebooks/analysis/data_raw_inspection.ipynb: 3
pandas == 1.1.4

# case-law-explorer/data_transformation/utils.py: 7
python_dateutil == 2.8.1

# case-law-explorer/data_extraction/caselaw/legal_intelligence/legal_intelligence_extractor.py: 10
# case-law-explorer/data_extraction/citations/rechtspraak_citations_extractor_incremental.py: 14
# case-law-explorer/definitions/storage_handler.py: 7
python_dotenv == 0.15.0

# case-law-explorer/data_extraction/citations/rechtspraak_citations_extractor_incremental.py: 4
rdflib == 6.1.1

# case-law-explorer/data_extraction/caselaw/echr/ECHR_metadata_harvester.py: 2
# case-law-explorer/data_extraction/caselaw/legal_intelligence/legal_intelligence_extractor.py: 4
# case-law-explorer/data_extraction/citations/rechtspraak_citations_extractor_incremental.py: 1
# case-law-explorer/notebooks/ECHR_metadata_harvester.ipynb: 1
requests == 2.26.0

# case-law-explorer/data_loading/clients/opensearch.py: 5
requests_aws4auth == 1.1.1

# case-law-explorer/network_analysis/correlation_calculator: 5
matplotlib == 2.5.1

# case-law-explorer/data_extraction/caselaw/echr: 6-11
Selenium == 4.4.3